WIP: Merge Supabase Vector Store nodes
