+++ packages/nodes-base/credentials/SupabaseApi.credentials.ts
@@ -25,6 +25,9 @@ export class SupabaseApi implements ICredentialType {
 			name: 'serviceRole',
 			type: 'string',
 			default: '',
+			typeOptions: {
+				password: true,
+			},
 		},
 	];
